turned ai_assistant_refactor in ai_assistant
